+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - `pyrotorque`: `QueueManager` now has a default `matcher` setting of `is_complete=no is_ignored=no is_open=no message="" done=0`
+- `pyrotorque`: The `args` paremeter for `pyrosimple.job.action:Action` now accepts non-string values.
 
 ## [2.11.2] - 2023-09-20